Atualuzar dependencias, migrations e factories

+++ database/migrations/2020_09_19_234828_create_avaliacaos_table.php
@@ -17,9 +17,9 @@ class CreateAvaliacaosTable extends Migration
             $table->id();
             $table->float('nota');
             $table->text('texto');
-            $table->integer('cliente_id')->unsigned();
+            $table->bigInteger('cliente_id')->unsigned();
             $table->foreign('cliente_id')->references('id')->on('users');
-            $table->integer('vendedor_id')->unsigned();
+            $table->bigInteger('vendedor_id')->unsigned();
             $table->foreign('vendedor_id')->references('id')->on('users');
             $table->timestamps();
         });
