+++ database/migrations/2020_09_20_020525_create_categorias_table.php
@@ -17,9 +17,9 @@ class CreateCategoriasTable extends Migration
             $table->id();
             $table->timestamps();
             $table->string('nome');
-            $table->integer('administrador_id')->unsigned();
+            $table->bigInteger('administrador_id')->unsigned();
             $table->foreign('administrador_id')->references('id')->on('users');
-            $table->integer('categoria_id')->unsigned()->nullable();
+            $table->bigInteger('categoria_id')->unsigned()->nullable();
             $table->foreign('categoria_id')->references('id')->on('categorias');
         });
     }
